routellm/middleware/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
pyyaml>=6.0

# For web UI
gradio>=3.50.0

# For embeddings
litellm>=1.0.0
